Move project module to api, reduce cli & app usage in other modules
