+++ src/snowcli/api/project/definition_manager.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import List, Optional
 
 from snowcli.api.exceptions import MissingConfiguration
-from snowcli.cli.project.definition import load_project_definition
+from snowcli.api.project.definition import load_project_definition
 
 
 def _compat_is_mount(path: Path):
