+++ src/snowcli/cli/nativeapp/run_processor.py
@@ -7,6 +7,7 @@ import typer
 from click import ClickException
 from rich import print
 from snowcli.api.exceptions import SnowflakeSQLExecutionError
+from snowcli.api.project.util import unquote_identifier
 from snowcli.cli.nativeapp.constants import (
     COMMENT_COL,
     INTERNAL_DISTRIBUTION,
@@ -30,7 +31,6 @@ from snowcli.cli.nativeapp.policy import PolicyBase
 from snowcli.cli.nativeapp.utils import find_first_row
 from snowcli.cli.object.stage.diff import DiffResult
 from snowcli.cli.object.stage.manager import StageManager
-from snowcli.cli.project.util import unquote_identifier
 from snowflake.connector import ProgrammingError
 from snowflake.connector.cursor import DictCursor, SnowflakeCursor
