+++ tests/nativeapp/test_package_scripts.py
@@ -7,7 +7,7 @@ from snowcli.cli.nativeapp.exceptions import (
     MissingPackageScriptError,
 )
 from snowcli.cli.nativeapp.run_processor import NativeAppRunProcessor
-from snowcli.cli.project.definition_manager import DefinitionManager
+from snowcli.api.project.definition_manager import DefinitionManager
 from snowflake.connector import ProgrammingError
 
 from tests.nativeapp.patch_utils import mock_connection
